Code refactor. Also include rubocop gem to encourage consistent code throughout the project and remove unused code in spec_helper.
