+++ lib/fork_break/breakpoints.rb
@@ -1,0 +1,7 @@
+module ForkBreak
+  module Breakpoints
+    def breakpoints
+      ForkBreak::Process.breakpoint_setter
+    end
+  end
+end
