+++ lib/fork_break/version.rb
@@ -1,0 +1,3 @@
+module ForkBreak
+  VERSION = '0.1.2'.freeze
+end
