+++ spec/spec_helper.rb
@@ -1,12 +1,6 @@
-require 'rubygems'
-require 'bundler/setup'
-require 'fork_break' # and any other gems you need
+require_relative '../lib/fork_break'
 
 RSpec.configure do |config|
-  config.treat_symbols_as_metadata_keys_with_true_values = true
-  config.run_all_when_everything_filtered = true
-  config.filter_run :focus
-
   # Run specs in random order to surface order dependencies. If you find an
   # order dependency and want to debug it, you can fix the order by providing
   # the seed, which is printed after each run.
